% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature_list.R
\name{create_feature_list}
\alias{create_feature_list}
\title{Create a Feature List from Raster Files}
\usage{
create_feature_list(
  filename = "feature_list.txt",
  output_dir = NULL,
  weight = NULL,
  wgrp = NULL,
  condition = NULL,
  group = NULL,
  retention = NULL,
  threshold = NULL,
  spp_file_dir,
  recursive = FALSE,
  spp_file_pattern = ".+\\\\.(tif|img)$"
)
}
\arguments{
\item{filename}{A character string specifying the name of the output file.
The file will have a \code{.txt} extension if not provided.}

\item{output_dir}{A character string specifying the output directory where the
feature list file should be saved. If NULL, the file will be saved
in the current working directory.}

\item{weight}{An optional numeric vector (\code{float}) to assign weights to the features in the list.}

\item{wgrp}{An optional integer vector (\code{int}) specifying a weight group number.}

\item{condition}{An optional integer vector (\code{int}) representing the condition transform number
associated with the rasters.}

\item{group}{An optional integer vector (\code{int}) representing the output group number for each raster.}

\item{retention}{An optional integer vector (\code{int}) representing the retention transform number for each raster.}

\item{threshold}{An optional numeric vector (\code{float}) representing the threshold value below which
input files are cut to zero.}

\item{spp_file_dir}{A character string specifying the directory containing the raster files.}

\item{recursive}{A logical value indicating whether to search subdirectories of \code{spp_file_dir}
for raster files. Default is FALSE.}

\item{spp_file_pattern}{A regular expression pattern used to filter the raster files
matching \code{.tif} and \code{.img} files.}
}
\value{
A text file containing a feature list of rasters along with any additional attributes
specified by the user. The file is saved to the location specified by \code{filename}
in \code{output_dir}.
}
\description{
This function generates a feature list from raster files in a specified directory,
adding optional attributes to the list based on user-defined parameters. The resulting
feature list is written to a text file.
}
\examples{
\dontrun{
create_feature_list(spp_file_dir = "path/to/raster/files",
                    output_dir = "path/to/output",
                    weight = c(0.1, 0.2, 0.3),
                    threshold = 0.5)
}
}
