% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_settings_file.R
\name{create_settings_file}
\alias{create_settings_file}
\title{Create a Settings File for Zonation 5 Analysis}
\usage{
create_settings_file(
  file_name,
  zero_mode = NULL,
  feature_list_file = NULL,
  external_solution_file = NULL,
  analysis_area_mask_layer = NULL,
  weight_groups_file = NULL,
  condition_link_file = NULL,
  hierarchic_mask_layer = NULL,
  cost_layer = NULL,
  retention_link_file = NULL
)
}
\arguments{
\item{file_name}{A character string specifying the name and/or location of the output settings file.
The file will have a \code{.z5} extension if not provided.}

\item{zero_mode}{Optional. A character string specifying the zero mode parameter for Zonation.}

\item{feature_list_file}{A character string specifying the full path and/or name of the feature list file.
This is a compulsory parameter.}

\item{external_solution_file}{A character string specifying the full path and/or name of the external solution file.}

\item{analysis_area_mask_layer}{A character string specifying the full path and/or name of the analysis area mask layer file.}

\item{weight_groups_file}{A character string specifying the full path and/or name of the weight groups file.}

\item{condition_link_file}{A character string specifying the full path and/or name of the condition link file.}

\item{hierarchic_mask_layer}{A character string specifying the full path and/or name of the hierarchic mask layer file.}

\item{cost_layer}{A character string specifying the full path and/or name of the cost layer file.}

\item{retention_link_file}{A character string specifying the full path and/or name of the retention link file.}
}
\value{
A \code{.z5} file containing the specified settings, saved to the location specified by \code{file_name}.
}
\description{
This function generates a settings file with various parameters related to
the input data and analysis configuration. The resulting settings
file is saved with a \code{.z5} extension and can be used directly in the Zonation 5 software.
}
\examples{
\dontrun{
create_settings_file(file_name = "settings.z5",
                     feature_list_file = "path/to/feature_list.txt",
                     external_solution_file = "path/to/solution.txt",
                     analysis_area_mask_layer = "path/to/mask_layer.tif")
}
}
