% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_list.R
\name{feature_list}
\alias{feature_list}
\title{Create a feature list from raster files}
\usage{
feature_list(spp_file_dir, weight = NULL, group = NULL, threshold = NULL)
}
\arguments{
\item{spp_file_dir}{A character string specifying the directory containing the raster files.}

\item{weight}{An optional numeric vector (\code{float}) to assign weights to the features in the list.}

\item{group}{An optional integer vector (\code{int}) representing the output group number for each raster.}

\item{threshold}{An optional numeric vector (\code{float}) representing the threshold value below which
input files are cut to zero.}
}
\value{
A text file containing a feature list of rasters along with any additional attributes
specified by the user.
}
\description{
This function generates a feature list from raster files in a specified directory,
adding optional attributes to the list based on user-defined parameters. The resulting
feature list is written to a text file. Supported raster formats include GeoTIFF
(.tif, .tiff), ERDAS Imagine (.img), and ASCII Grid (.asc).
}
\examples{
\dontrun{
feature_list(spp_file_dir = "path/to/raster/files")
}
}
