% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_command_file.R
\name{run_command_file}
\alias{run_command_file}
\title{Run a Zonation command file}
\usage{
run_command_file(folder)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing
the Zonation command file.}
}
\description{
This function runs a Zonation 5 analysis directly from R by executing a
command file located in the specified folder. On Windows, the function looks
for a single \code{.cmd} file; on Linux, it looks for a single \code{.sh} file.
}
\examples{
\dontrun{
run_command_file("C:/path/to/folder")
}
}
