% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_zonation5.R
\name{run_zonation5}
\alias{run_zonation5}
\title{Run Zonation 5}
\usage{
run_zonation5(folder)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing the \code{.cmd} file.}
}
\description{
This function runs a Zonation 5 analysis directly from R. If multiple .cmd files
are found in the folder, it automatically selects and runs the first one
}
\examples{
\dontrun{
run_zonation5("C:/path/to/folder")
}
}
