% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature_list.R
\name{create_feature_list}
\alias{create_feature_list}
\title{Create a Feature List from Raster Files}
\usage{
create_feature_list(
  weight = NULL,
  wgrp = NULL,
  condition = NULL,
  group = NULL,
  retention = NULL,
  threshold = NULL,
  spp_file_dir
)
}
\arguments{
\item{weight}{An optional numeric vector (\code{float}) to assign weights to the features in the list.}

\item{wgrp}{An optional integer vector (\code{int}) specifying a weight group number.}

\item{condition}{An optional integer vector (\code{int}) representing the condition transform number
associated with the rasters.}

\item{group}{An optional integer vector (\code{int}) representing the output group number for each raster.}

\item{retention}{An optional integer vector (\code{int}) representing the retention transform number for each raster.}

\item{threshold}{An optional numeric vector (\code{float}) representing the threshold value below which
input files are cut to zero.}

\item{spp_file_dir}{A character string specifying the directory containing the raster files.}
}
\value{
A text file containing a feature list of rasters along with any additional attributes
specified by the user.
}
\description{
This function generates a feature list from raster files in a specified directory,
adding optional attributes to the list based on user-defined parameters. The resulting
feature list is written to a text file.
}
\examples{
\dontrun{
create_feature_list(spp_file_dir = "path/to/raster/files")
}
}
