% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zonation5_call.R
\name{create_zonation5_call}
\alias{create_zonation5_call}
\title{Create a Zonation 5 Command File}
\usage{
create_zonation5_call(
  os = "Windows",
  zonation_path,
  flags = "",
  marginal_loss_mode,
  gui_activated = FALSE,
  settings_file,
  output_directory = "output",
  command_file = "command_file"
)
}
\arguments{
\item{os}{Operating system. Default is "Windows".}

\item{zonation_path}{The specification for the path where Zonation 5 is installed.}

\item{flags}{Flags that control which analytical options are used. Used to include single letter codes
that switch analytical options on.}

\item{marginal_loss_mode}{The marginal loss rule is specified using this parameter.}

\item{gui_activated}{This parameter controls whether the Graphical User Interface (GUI)
is launched when running the command file. The default is FALSE (GUI not activated).}

\item{settings_file}{Path to the settings file.}

\item{output_directory}{Output directory for results.}

\item{command_file}{Path and/or name to save the command file.}
}
\value{
The Zonation 5 command file.
}
\description{
This function generates a command file to run Zonation 5. It is also used to specify the analysis options,
settings, and output location. The file is saved with a \code{.cmd} (Windows) or \code{.sh} (Linux) suffix.
}
\examples{
\dontrun{
create_zonation5_call(os = "Windows",
                      zonation_path = "C:/Program Files (x86)/Zonation5",
                      marginal_loss_mode = "ABF",
                      settings_file = "settings.txt")

create_zonation5_call(os = "Linux",
                      zonation_path = "/usr/local/zonation",
                      marginal_loss_mode = "ABF",
                      gui_activated = TRUE)
}
}
