% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_command_file.R
\name{run_command_file}
\alias{run_command_file}
\title{Run Zonation 5}
\usage{
run_command_file(folder)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing the \code{.cmd} file.}
}
\description{
This function runs a Zonation 5 analysis directly from R by executing a \code{.cmd} file
located in the specified folder. If multiple \code{.cmd} files are present, it stops with an error,
requiring only one \code{.cmd} file to be in the folder.
}
\examples{
\dontrun{
run_command_file("C:/path/to/folder")
}
}
